Adds working with the server, loading and updating data
