+++ src/model/points-model.js
@@ -1,12 +1,13 @@
 import Observable from '../framework/observable.js';
 import { UpdateType } from '../const.js';
+import PointsAdapterService from '../service/points-adapter-service.js';
 
 export default class PointsModel extends Observable {
   #points = [];
-  #pointsApiSevrice = null;
   #offersModel = [];
   #destiationModel = [];
-
+  #pointsApiSevrice = null;
+  #pointsAdapterService = new PointsAdapterService();
 
   constructor({pointsApiService, offersModel, destinationsModel}) {
     super();
@@ -24,9 +25,11 @@ export default class PointsModel extends Observable {
       await this.#offersModel.init();
       await this.#destiationModel.init();
       const points = await this.#pointsApiSevrice.points;
-      this.#points = points.map(this.#adaptToClient);
+      this.#points = points.map(this.#pointsAdapterService.adaptToClient);
     } catch (error) {
       this.#points = [];
+      this.#offersModel = [];
+      this.#destiationModel = [];
     }
 
     this._notify(UpdateType.INIT);
@@ -41,7 +44,7 @@ export default class PointsModel extends Observable {
 
     try {
       const response = await this.#pointsApiSevrice.updatePoint(update);
-      const updatedPoint = this.#adaptToClient(response);
+      const updatedPoint = this.#pointsAdapterService.adaptToClient(response);
       this.#points = [
         ...this.#points.slice(0, index),
         updatedPoint,
@@ -75,21 +78,5 @@ export default class PointsModel extends Observable {
     ];
 
     this._notify(updateType);
-  }
-
-  #adaptToClient(point) {
-    const adaptedPoint = {
-      ...point,
-      basePrice: point['base_price'],
-      dateFrom: point['date_from'] !== null ? new Date(point['date_from']) : point['date_from'],
-      dateTo: point['date_to'] !== null ? new Date(point['date_to']) : point['date_to'],
-      isFavorite: point['is_favorite']
-    };
-
-    delete adaptedPoint['base_price'];
-    delete adaptedPoint['date_from'];
-    delete adaptedPoint['date_to'];
-    delete adaptedPoint['is_favorite'];
-    return adaptedPoint;
   }
 }
