--- src/points-api-service.js
@@ -1,50 +1,0 @@
-import ApiService from './framework/api-service.js';
-
-const Method = {
-  GET: 'GET',
-  PUT: 'PUT'
-};
-
-export default class PointsApiService extends ApiService {
-  get points() {
-    return this._load({url: 'big-trip/points'})
-      .then(ApiService.parseResponse);
-  }
-
-  get offers() {
-    return this._load({url: 'big-trip/offers'})
-      .then(ApiService.parseResponse);
-  }
-
-  get destinations() {
-    return this._load({url: 'big-trip/destinations'})
-      .then(ApiService.parseResponse);
-  }
-
-  async updatePoint(point) {
-    const response = await this._load({
-      url: `big-trip/points/${point.id}`,
-      method: Method.PUT,
-      body: JSON.stringify(this.#adaptToServer(point)),
-      headers: new Headers({'Content-Type': 'application/json'})
-    });
-
-    return await ApiService.parseResponse(response);
-  }
-
-  #adaptToServer(point) {
-    const adaptedPoint = {
-      ...point,
-      'base_price': point.basePrice,
-      'date_from': point.dateFrom instanceof Date ? point.dateFrom.toISOString() : null,
-      'date_to': point.dateTo instanceof Date ? point.dateTo.toISOString() : null,
-      'is_favorite': point.isFavorite
-    };
-
-    delete adaptedPoint.basePrice;
-    delete adaptedPoint.dateFrom;
-    delete adaptedPoint.dateTo;
-    delete adaptedPoint.isFavorite;
-    return adaptedPoint;
-  }
-}
